embedding-service/requirements.txt
flask
requests 

weaviate-client

docling

langchain-text-splitters
langchain-core
langchain-openai
langchain-google-genai
langchain-voyageai
langchain-anthropic